requirements.txt
pandas==0.20.3
scikit_image==0.13.0
matplotlib==2.0.2
numpy==1.13.1
Flask==0.12.2
Pillow==4.2.1
skimage==0.0
scikit_learn==0.19.0
gunicorn==18.0
